requirements.txt
flask
gunicorn
flask_socketio
psycopg2-binary